Refactor and modularize students journal application
Extracted and organized functionalities into separate modules for better readability and maintainability. This update includes splitting student data management, user input handling, and helper methods into dedicated files. Additionally, updated the README with a clear description of the application's features and commands.
