+++ lesson3/Jornal_interface/README_for_interface.md
@@ -1,7 +1,58 @@
-# Hillal_Python_pro
+# Students Journal Application
+
+
+Description
+
+Students Journal Application is a simple graphical user interface (GUI) application for 
+managing student data, their grades, and additional information. This application is built using the 
+PySimpleGUI library.
+
+Features
+
+View All Entries: 
+
+Allows you to view all student records along with their grades.
+
+Add New Student: Capability to add new students to the journal.
+
+Search Student by ID: Ability to find a student using their ID.
+
+Update Student Data: Modify details of an already existing student.
+
+Delete Student Data: Ability to delete a student's record.
+
+Display Student Database Information: Shows the total number of students and average score.
+
+Commands
+
+quit: terminate the application
+
+help: display information about commands
+
+show: show the list of all student entries
+
+add: add a new student
+
+retrieve: find a student by ID
+
+remove: delete student data
+
+change: update student information
+
+Installation and Running
+
+Ensure you have Python 3.12.3 and PySimpleGUI installed.
+
+Clone the repository or download the project files to your computer.
+
+Place the journal_demo.png image in the application's directory.
+
+Run the script with the command main()
 
 For Using, need to install this package:
 
 pip install PySimpleGUI
 
 pip install tkintertable
+
+Hyllal Pyton_Pro
